% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_af_threshold_table.R
\name{adjust_af_threshold_table}
\alias{adjust_af_threshold_table}
\title{adjust_af_threshold_table}
\usage{
adjust_af_threshold_table(controls_dir = file.path(outdir, "Controls"),
  pbem_dir = file.path(outdir, "BaseErrorModel"),
  detection.specificity = 0.995, replicas = 10,
  replicas.in.parallel = 1, coverage_binning = 50)
}
\arguments{
\item{controls_dir}{folder with afth data. default: "outdir/Controls"}

\item{pbem_dir}{folder with pbem data. default: "outdir/BaseErrorModel"}

\item{detection.specificity}{default: 0.995}

\item{replicas}{default: 10}

\item{replicas.in.parallel}{default: 1}

\item{coverage_binning}{Bins of coverage into which divide allelic fractions. default: 50}
}
\value{
corrected af threshold table
}
\description{
adjust_af_threshold_table
}
