% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2positions.R
\name{bed2positions}
\alias{bed2positions}
\title{bed2positions}
\usage{
bed2positions(targetbed)
}
\arguments{
\item{targetbed}{Genomic regions in the BED tab-delimited format.}
}
\value{
PosByChrom A list of data.frames, each for each chromosome.

chromosomes The chromosomes included in the BED file.
}
\description{
bed2positions
}
