% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bombanel_afs.R
\docType{data}
\name{bombanel_afs}
\alias{bombanel_afs}
\title{bombanel_afs}
\format{An object of class \code{numeric} of length 218.}
\usage{
bombanel_afs
}
\description{
Allelic Fractions from simulated data
}
\keyword{datasets}
