% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bombanel_tab_cov_pbem.R
\docType{data}
\name{bombanel_tab_cov_pbem}
\alias{bombanel_tab_cov_pbem}
\title{bombanel_tab_cov_pbem}
\format{An object of class \code{matrix} with 1999 rows and 218 columns.}
\usage{
bombanel_tab_cov_pbem
}
\description{
Matrix coverage_levels x pbem from simulated data
}
\keyword{datasets}
