% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_afthreshold.R
\name{compute_afthreshold}
\alias{compute_afthreshold}
\title{Compute Allelic Fraction thresholds}
\usage{
compute_afthreshold(outdir, pbem_dir = file.path(outdir,
  "BaseErrorModel"), outdir.afth.name = "Controls",
  coverage_binning = 50, probs = seq(0.9, 1, 1e-04))
}
\arguments{
\item{outdir}{output folder for this step analysis}

\item{pbem_dir}{folder with pbem data. default: file.path(outdir, "BaseErrorModel")}

\item{outdir.afth.name}{folder will be created in outdir. default: "Controls"}

\item{coverage_binning}{Bins of coverage into which divide allelic fractions. default: 50}

\item{probs}{quatiles to compute}
}
\value{
allelic fraction thresholds
}
\description{
Compute Allelic Fraction thresholds
}
\examples{
outdir <- tempdir()
pbem_dir <- system.file("extdata", "BaseErrorModel", package = "abemus")
outafth <- compute_afthreshold(outdir = outdir,pbem_dir = pbem_dir)
}
