% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cov_bins.R
\name{define_cov_bins}
\alias{define_cov_bins}
\title{Create bins of coverage}
\usage{
define_cov_bins(coverage_binning, min.coverage = 0,
  max.coverage = 5000)
}
\arguments{
\item{coverage_binning}{bin}

\item{min.coverage}{min coverage. default: 5000}

\item{max.coverage}{max coverage. default: 0}
}
\value{
list with covbin and lev
}
\description{
Create bins of coverage to stratify AFs
}
