% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{filter}
\usage{
filter(i, chromosomes, patient_folder, plasma.folder, germline.folder,
  name.patient, name.plasma, name.germline, pbem_dir, out1, out2, out3,
  mincov, minalt, mincovgerm, maxafgerm, AFbycov, minaf_cov_corrected, xbg,
  tab_cov_pbem, afs, covs, coverage_binning, njobs = 1)
}
\arguments{
\item{coverage_binning}{Bins of coverage into which divide allelic fractions. default: 50}
}
\description{
filter
}
