% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromListToDF.R
\name{fromListToDF}
\alias{fromListToDF}
\title{fromListToDF}
\usage{
fromListToDF(inputList)
}
\arguments{
\item{inputlist}{list to merge}
}
\value{
data.frame
}
\description{
fromListToDF
}
\author{
Davide Prandi
}
