% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_germlineset.R
\name{get_germlineset}
\alias{get_germlineset}
\title{Get list of pileups for germline samples}
\usage{
get_germlineset(main_sif, pacbamfolder, chrom)
}
\arguments{
\item{main_sif}{sample info file}

\item{pacbamfolder}{folder with pileups}

\item{chrom}{chromosome}
}
\value{
A character string with the path to germline pileups
}
\description{
Get list of pileups for germline samples
}
