% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbem.R
\name{get_pbem}
\alias{get_pbem}
\title{get_pbem}
\usage{
get_pbem(id, tgf)
}
\arguments{
\item{id}{index}

\item{tgf}{targeted data.frame}
}
\value{
row with bperr and tot_reads_supporting_alt
}
\description{
get_pbem
}
