% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats_pbem.R
\name{get_stats_pbem}
\alias{get_stats_pbem}
\title{get pbem statistics}
\usage{
get_stats_pbem(dt, n_pos_af_th)
}
\arguments{
\item{dt}{completetab_dbsnp as data.table}

\item{n_pos_af_th}{When compute pbem, count in how many germline samples the position has an AF >= n_pos_af_th. default: 0.2}
}
\value{
out data.table
}
\description{
get pbem statistics
}
