% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovaf.R
\name{getCovaf}
\alias{getCovaf}
\title{Compute the per-base error model}
\usage{
getCovaf(data, max.vaf = 0.2, min.cov = 10)
}
\arguments{
\item{data}{The output from \code{importData()} function.}

\item{max.vaf}{Threshold on VAF. default: 0.2}

\item{min.cov}{Threshold on coverage. default: 10}
}
\value{
A list with coverage (cov) and VAF (vaf) values split by chromosome.
}
\description{
Compute the per-base error model
}
