% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sif.R
\name{import_sif}
\alias{import_sif}
\title{Read and check the sample info file}
\usage{
import_sif(filesif)
}
\arguments{
\item{filesif}{A tab delimeted file without header where columns are:}
}
\value{
A list containing 2 data frames:
[[1]] Table with unique germline samples (use for GSE distribution and pbem computation steps)
[[2]] Table with only case samples having a matched control sample (use for call snvs step)
}
\description{
Read and check the sample info file
}
