% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{Import and process data necessary for abemus workflow.}
\usage{
importData(samples.info.file, pacbam)
}
\arguments{
\item{samples.info.file}{A tab-delimited two columns file listing case and control samples.}

\item{pacbam}{Path to folder with data generated by PaCBAM.}
}
\value{
A list with following objects:
}
\description{
Import and process data necessary for abemus workflow.
}
