% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos2bperr.R
\name{pos2bperr}
\alias{pos2bperr}
\title{pos2bperr}
\usage{
pos2bperr(id, targets, germlineset, step, chrom, lev, covbin,
  af_max_to_compute_thresholds, coverage_min_to_compute_thresholds,
  af_max_to_compute_pbem, coverage_min_to_compute_pbem, n_pos_af_th)
}
\arguments{
\item{id}{index}

\item{targets}{targeted positions}

\item{germlineset}{unique list of germlines samples from sif, see get_germlineset(sif)}

\item{step}{into how many positions to split the chrom file. default: 5000}

\item{chrom}{chromosome to process}

\item{lev}{Levels of coverage, see define_cov_bins(coverage_binning)}

\item{covbin}{Bins of coverage, see define_cov_bins(coverage_binning)}

\item{af_max_to_compute_thresholds}{To compute AF thresholds, consider only positions with AF <= af_max_to_compute_thresholds. default 0.2}

\item{coverage_min_to_compute_thresholds}{To compute AF threshold, consider only positions with coverage >= coverage_min_to_compute_thresholds. default 10}

\item{af_max_to_compute_pbem}{To compute pbem, consider only positions with AF <= af_max_to_compute_pbem. default: 0.2}

\item{coverage_min_to_compute_pbem}{To compute pbem, consider only positions with coverage >= coverage_min_to_compute_pbem. default: 10}

\item{n_pos_af_th}{When compute pbem, count in how many germline samples the position has an AF >= n_pos_af_th. default: 0.2}
}
\description{
Compute the per-base error model for each position
}
