% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantilezaf.R
\name{quantile.zaf}
\alias{quantile.zaf}
\title{quantile.zaf}
\usage{
\method{quantile}{zaf}(x, probs = seq(0, 1, 0.25), nz)
}
\arguments{
\item{x}{vafcov[idx,1]}

\item{probs}{quantiles}

\item{nz}{zeros}
}
\value{
quantiles
}
\description{
quantile.zaf
}
