% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pacbam_bychrom.R
\name{split_pacbam_bychrom}
\alias{split_pacbam_bychrom}
\title{Split PaCBAM ouputs by chromosomes}
\usage{
split_pacbam_bychrom(targetbed, pacbamfolder, mc.cores = 1,
  pacbamfolder_bychrom)
}
\arguments{
\item{targetbed}{Target regions in BED format}

\item{pacbamfolder}{Folder with original outputs generated by PaCBAM. This folder should contain both .pabs and .pileups files.}

\item{mc.cores}{Process n chromosomes in parallel. default = 1}

\item{pacbamfolder_bychrom}{Output folder in which subfolders for each sample will be created.}
}
\description{
Split PaCBAM ouputs by chromosomes
}
\examples{
targetbed <- system.file("extdata", "regions_toy.bed", package = "abemus")
pacbamfolder <- system.file("extdata", "pacbam_data", package = "abemus")
split_pacbam_bychrom(targetbed=targetbed,pacbamfolder=pacbamfolder,pacbamfolder_bychrom=tempdir())
}
