% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_optimal_R.R
\docType{data}
\name{tab_optimal_R}
\alias{tab_optimal_R}
\title{[data] Summary table to select the best scaling factor R}
\format{An object of class \code{data.frame} with 40 rows and 3 columns.}
\usage{
tab_optimal_R
}
\description{
[data] Summary table to select the best scaling factor R
}
\keyword{datasets}
