% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_af_threshold_table.R
\name{adjust_af_threshold_table}
\alias{adjust_af_threshold_table}
\title{adjust_af_threshold_table}
\usage{
adjust_af_threshold_table(controls_dir, pbem_dir, detection.specificity,
  replicas, replicas.in.parallel, coverage_binning, coeffvar.threshold)
}
\arguments{
\item{controls_dir}{folder with afth data. default: file.path(outdir,"Controls")}

\item{pbem_dir}{folder with pbem data. default: file.path(outdir,"BaseErrorModel")}

\item{detection.specificity}{default: 0.995}

\item{replicas}{default: 1000}

\item{replicas.in.parallel}{default: 1}

\item{coverage_binning}{Bins of coverage into which divide allelic fractions. default: 50}

\item{coeffvar.threshold}{Consider a bin as stable if the Coeff of variations after n replicas is lower than coeffvar.threshold. default: 0.01}
}
\value{
corrected af threshold table
}
\description{
adjust_af_threshold_table
}
