% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_scaling_factor.R
\name{apply_scaling_factor}
\alias{apply_scaling_factor}
\title{apply_scaling_factor}
\usage{
apply_scaling_factor(tabindex, R = 1)
}
\arguments{
\item{tabindex}{output from callsnvs()}

\item{R}{scaling factor to be applied at filter.pbem_coverage value. default = 1}
}
\description{
apply_scaling_factor
}
