% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2positions.R
\name{bed2positions}
\alias{bed2positions}
\title{From BED format file to a list of loci}
\usage{
bed2positions(targetbed, chrom_to_extract, get_only_chromosomes = F)
}
\arguments{
\item{targetbed}{Genomic regions in the BED tab-delimited format.}

\item{chrom_to_extract}{the chromosome name to extract as i.e. "chr1" or "1" accordingly with BED file annotation}

\item{get_only_chromosomes}{return only the list of chromosomes in the BED file. default: FALSE}
}
\value{
PosByChrom A list of data.frames, each for each chromosome.

chromosomes The chromosomes included in the BED file.
}
\description{
From BED format file to a list of loci
}
