% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bombanel_covs.R
\docType{data}
\name{bombanel_covs}
\alias{bombanel_covs}
\title{[data] Coverage levels from simulated data}
\format{
An object of class \code{numeric} of length 1999.
}
\usage{
bombanel_covs
}
\description{
[data] Coverage levels from simulated data
}
\keyword{datasets}
