% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callsnvs.R
\name{callsnvs}
\alias{callsnvs}
\title{Call somatic SNVs by using both global and local error models}
\usage{
callsnvs(sample.info.file, targetbed, pacbamfolder_bychrom,
  pbem_dir = file.path(outdir, "BaseErrorModel"),
  controls_dir = file.path(outdir, "Controls"),
  detection.specificity = 0.995, replicas = 1000,
  replicas.in.parallel = 1, coeffvar.threshold = 0.01,
  AFbycov = TRUE, mincov = 10, mincovgerm = 10, minalt = 1,
  maxafgerm = 0.2, coverage_binning = 50, outdir,
  outdir.calls.name = "Results", chrom.in.parallel = 1)
}
\arguments{
\item{sample.info.file}{sample info file listing cases and controls. tab-delimeted file}

\item{targetbed}{targeted regions in BED format.}

\item{pacbamfolder_bychrom}{folder with pileups}

\item{pbem_dir}{folder with pbem data. default: file.path(outdir, "BaseErrorModel")}

\item{controls_dir}{folder with afth data. default: file.path(outdir,"Controls")}

\item{detection.specificity}{The quantile of the GSE distribution(s) to use to compute the afth. default: 0.995}

\item{replicas}{Relpica sampling to define stability of afth in bins of coverage. default: 1000}

\item{replicas.in.parallel}{default: 1}

\item{coeffvar.threshold}{Consider a bin as stable if the Coeff of variations after n replicas is lower than coeffvar.threshold. default: 0.01}

\item{AFbycov}{Apply afth coverage based. default: AFbycov=TRUE}

\item{mincov}{Minimum locus coverage in case sample. default: 10}

\item{mincovgerm}{Minimum locus coverage in germline sample. default: 10}

\item{minalt}{Minimum number of reads supporting the alternative allele in case sample. default: 1}

\item{maxafgerm}{Maximum allelic fraction observed in matched germline locus. default: 0.2}

\item{coverage_binning}{Bins of coverage into which divide allelic fractions. default: 50}

\item{outdir}{output folder for this step analysis}

\item{outdir.calls.name}{folder will be created in outdir. default: "Results"}

\item{chrom.in.parallel}{number of chromosomes to run in parallel. default: 1}
}
\value{
list of data frames with path to files with SNV calls.
}
\description{
Call somatic SNVs by using both global and local error models
}
\examples{
sample.info.file <- system.file("extdata", "test_sif_toy.tsv", package = "abemus")
outdir <- tempdir()
targetbed <- system.file("extdata", "regions_toy.bed", package = "abemus")
pacbamfolder_bychrom <- system.file("extdata", "pacbam_data_bychrom", package = "abemus")
pbem_dir <- system.file("extdata", "BaseErrorModel", package = "abemus")
controls_dir <- system.file("extdata", "Controls", package = "abemus")
calls <- callsnvs(sample.info.file = sample.info.file,outdir=outdir,targetbed=targetbed,pbem_dir=pbem_dir,controls_dir=controls_dir,pacbamfolder_bychrom=pacbamfolder_bychrom,replicas = 2)
}
