% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pbem.R
\name{compute_pbem}
\alias{compute_pbem}
\title{compute_pbem}
\usage{
compute_pbem(sample.info.file, targetbed, pacbamfolder_bychrom, outdir,
  outdir.bperr.name = "BaseErrorModel", coverage_binning = 50,
  af_max_to_compute_thresholds = 0.2,
  coverage_min_to_compute_thresholds = 10,
  af_max_to_compute_pbem = 0.2, coverage_min_to_compute_pbem = 10,
  n_pos_af_th = 0.2, mc.cores = 1, step = 5000)
}
\arguments{
\item{sample.info.file}{sample info file listing cases and controls. tab-delimeted file}

\item{targetbed}{targeted regions in BED format.}

\item{pacbamfolder_bychrom}{folder with pileups}

\item{outdir}{output folder for this step analysis}

\item{outdir.bperr.name}{folder will be created in outdir. default: "BaseErrorModel"}

\item{coverage_binning}{Bins of coverage into which divide allelic fractions. default: 50}

\item{af_max_to_compute_thresholds}{To compute AF thresholds, consider only positions with AF <= af_max_to_compute_thresholds. default 0.2}

\item{coverage_min_to_compute_thresholds}{To compute AF threshold, consider only positions with coverage >= coverage_min_to_compute_thresholds. default 10}

\item{af_max_to_compute_pbem}{To compute pbem, consider only positions with AF <= af_max_to_compute_pbem. default: 0.2}

\item{coverage_min_to_compute_pbem}{To compute pbem, consider only positions with coverage >= coverage_min_to_compute_pbem. default: 10}

\item{n_pos_af_th}{When compute pbem, count in how many germline samples the position has an AF >= n_pos_af_th. default: 0.2}

\item{mc.cores}{mc.core param from mclapply. default: 1}

\item{step}{into how many positions to split the chrom file. default: 5000}
}
\value{
list(bperr, bperr_summary, bperr_tabstat)
}
\description{
Compute per-base error model on each targeted position and save AFs by bins of coverage
}
