% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_case_mean_coverage.R
\name{get_case_mean_coverage}
\alias{get_case_mean_coverage}
\title{get_case_mean_coverage}
\usage{
get_case_mean_coverage(tabindex, pacbamfolder_bychrom)
}
\arguments{
\item{tabindex}{output calls$tabsnvs_index from callsnvs() function.}

\item{pacbamfolder_bychrom}{folder with pileups split by chromosome.}
}
\value{
sample.info.file with 1 additional column reporting the mean coverage for each case having a matched-control
}
\description{
get_case_mean_coverage
}
