% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_case_mean_coverage.R
\name{get_case_mean_coverage}
\alias{get_case_mean_coverage}
\title{Compute the mean coverage of a case sample}
\usage{
get_case_mean_coverage(tabindex, pacbamfolder_bychrom)
}
\arguments{
\item{tabindex}{output calls$tabsnvs_index from callsnvs() function.}

\item{pacbamfolder_bychrom}{folder with pileups split by chromosome.}
}
\value{
sample.info.file with 1 additional column reporting the mean coverage for each case having a matched-control
}
\description{
Compute the mean coverage of a case sample
}
\examples{
sample.info.file <- system.file("extdata", "test_sif_toy.tsv", package = "abemus")
outdir <- tempdir()
targetbed <- system.file("extdata", "regions_toy.bed", package = "abemus")
pacbamfolder_bychrom <- system.file("extdata", "pacbam_data_bychrom", package = "abemus")
pbem_dir <- system.file("extdata", "BaseErrorModel", package = "abemus")
controls_dir <- system.file("extdata", "Controls", package = "abemus")
m<-callsnvs(sample.info.file,outdir,targetbed,pbem_dir,controls_dir,pacbamfolder_bychrom,replicas=1)
tabindex <- m$tabsnvs_index
tabindex <- get_case_mean_coverage(tabindex = tabindex,pacbamfolder_bychrom = pacbamfolder_bychrom)
}
