% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_size.R
\name{get_target_size}
\alias{get_target_size}
\title{get_target_size}
\usage{
get_target_size(targetbed, Mbp = TRUE)
}
\arguments{
\item{targetbed}{targeted regions in BED format}

\item{Mbp}{return count as Mbp. default: TRUE}
}
\value{
Mbp covered in the BED file
}
\description{
get_target_size
}
