% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPbem.R
\name{getPbem}
\alias{getPbem}
\title{Compute the per-base error model}
\usage{
getPbem(data, max.vaf.pbem = 0.2, cov.min.pbem = 10)
}
\arguments{
\item{data}{The output from \code{importData()} function.}

\item{max.vaf.pbem}{Threshold on VAF. default: 0.2}

\item{cov.min.pbem}{Threshold on coverage. default: 10}
}
\value{
A list with pbem values split by chromosome.
}
\description{
Compute the per-base error model
}
